# services/gateway/docs/swagger.yaml
basePath: /
definitions:
  ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  authResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/userResponse'
    type: object
...
  emailRequest:
    properties:
      email:
        type: string
    type: object
...
  loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  mediaResponse:
    properties:
      checksum:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      id:
        type: string
      media_type:
        type: string
      original_filename:
        type: string
      size:
        type: integer
      uploaded_by:
        type: string
    type: object
...
  messageResponse:
    properties:
      message:
        type: string
    type: object
...
  permissionsResponse:
    properties:
      permissions:
        items:
          type: string
        type: array
    type: object
...
  refreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  registerRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  resetPasswordRequest:
    properties:
      new_password:
        type: string
      token:
        type: string
    type: object
...
  tokenRequest:
    properties:
      token:
        type: string
    type: object
...
  userResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      last_seen_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Единый публичный HTTP-контракт Go-бэкенда.
  title: General Project API
  version: "1.0"
paths:
  /v1/auth/confirm-registration:
    post:
      consumes:
      - application/json
      parameters:
      - description: Confirmation-токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Подтвердить регистрацию
      tags:
      - auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Вход пользователя
      tags:
      - auth
  /v1/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh-токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/refreshRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Завершить сессию
      tags:
      - auth
  /v1/auth/me/permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permissionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить permissions пользователя
      tags:
      - auth
  /v1/auth/password-reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset-токен и новый пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Сбросить пароль
      tags:
      - auth
  /v1/auth/password-reset-requests:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/emailRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/messageResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Запросить сброс пароля
      tags:
      - auth
  /v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh-токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Обновить токены
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/registerRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /v1/auth/registration-confirmation-requests:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/emailRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/messageResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Повторно отправить подтверждение регистрации
      tags:
      - auth
...
  /v1/media:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Файл
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/mediaResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Загрузить медиа
      tags:
      - media
  /v1/media/{mediaID}:
    delete:
      parameters:
      - description: UUID медиаобъекта
        in: path
        name: mediaID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Удалить медиа
      tags:
      - media
    get:
      parameters:
      - description: UUID медиаобъекта
        in: path
        name: mediaID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mediaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Получить метаданные медиа
      tags:
      - media
...
swagger: "2.0"